folder restructure and added sass
