+++ ui/package.json
@@ -3,6 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "node-sass-chokidar": "~1.3.3",
+    "npm-run-all": "^4.1.3",
     "patternfly": "^3.52.2",
     "patternfly-react": "^2.11.1",
     "react": "^16.4.1",
@@ -11,10 +13,14 @@
     "react-scripts": "1.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
+    "start-js": "react-scripts start",
+    "start": "npm-run-all -p watch-css start-js",
+    "build-js": "react-scripts build",
+    "build": "npm-run-all build-css build-js",
     "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "build-css": "node-sass-chokidar --include-path ./node_modules/patternfly/dist/sass --include-path ./node_modules/patternfly/node_modules/bootstrap-sass/assets/stylesheets --include-path ./node_modules/patternfly/node_modules/font-awesome-sass/assets/stylesheets src/ -o src/",
+    "watch-css": "npm run build-css && node-sass-chokidar --include-path ./node_modules/patternfly/dist/sass --include-path ./node_modules/patternfly/node_modules/bootstrap-sass/assets/stylesheets --include-path ./node_modules/patternfly/node_modules/font-awesome-sass/assets/stylesheets src/ -o src/ --watch --recursive"
   },
   "homepage": "/"
 }
