+++ ui/src/components/App.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router, Switch, Redirect, Route } from 'react-router-d
 import Overview from './Overview';
 import Client from './Client';
 
-import './App.css';
+import './App.css'
 
 class App extends Component {
   render() {
@@ -12,7 +12,7 @@ class App extends Component {
         <div className="App">
           <Switch>
             <Route exact path="/overview" component={Overview}/>
-            <Route exact path="/mobileclient/:id" component={Client}/>
+            <Route exact path="/mobileclien/:id" component={Client}/>
 
             {/* Default redirect */}
             <Redirect to="/overview" />
