+++ ui/src/components/App.scss
@@ -1,0 +1,2 @@
+@import "../style/variables";
+@import "patternfly";
