--- ui/src/index.css
@@ -1,2 +1,0 @@
-@import url("~patternfly/dist/css/patternfly.min.css");
-@import url("~patternfly/dist/css/patternfly-additions.min.css");
