+++ ui/src/style/_variables.scss
@@ -1,0 +1,3 @@
+$font-path:       "~patternfly/dist/fonts/";
+$img-path:        "~patternfly/dist/img/";
+$icon-font-path:  "~patternfly/node_modules/bootstrap-sass/assets/fonts/bootstrap/";
